requirements.txt
requests==2.30.0
bilibili-api-python==15.4.4
PySide6==6.5.1